pull out processeventconsumer
