+++ comp/system-probe/eventmonitor/def/component.go
@@ -8,7 +8,6 @@ package eventmonitor
 
 import (
 	"github.com/DataDog/datadog-agent/comp/system-probe/types"
-	"github.com/DataDog/datadog-agent/pkg/eventmonitor/consumers"
 )
 
 // team: agent-security
@@ -16,12 +15,4 @@ import (
 // Component is the component type.
 type Component interface {
 	types.SystemProbeModuleComponent
-}
-
-type ProcessEventConsumerComponent interface {
-	ID() string
-	ChanSize() int
-	EventTypes() []consumers.ProcessConsumerEventTypes
-	Set(*consumers.ProcessConsumer)
-	Get() *consumers.ProcessConsumer
 }
